% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{cpm}
\alias{cpm}
\title{Calculate counts per million (cpm)}
\usage{
cpm(mat)
}
\arguments{
\item{mat}{data frame count matrix with rows as genes and columns as samples}
}
\value{
data frame with CPM normalized gene counts
}
\description{
Function to calculate counts per million (CPM) using a gene count matrix
}
